Update patch set 3
Patch Set 3: Code-Review+1

Changed the audio file to /system/etc/sound/GT-N8013.

The minor issue we face is that if ro.product.device is set to GT-N801x, people will be unable to flash other ROMs because the asserts will fail. So I've left the ro.product.device as GT-N8013, but audio_hw.c depends on this to choose what sound configuration file to load. The easiest thing is to just change the name of the file.

This still feels slightly awkward to me, because ro.product.model is still set as GT-N801x. Should we change that too for consistency? The main issue is that I don't want to lock people into a ROM because we're using a different device name; that's just silly (yes, they can always remove the asserts, but that's also silly to have to do).

Patch-set: 3
Label: Code-Review=+1