Update patch set 4
Patch Set 4:

The asserts are fine:

(getprop("ro.product.device") == "c0" || getprop("ro.build.product") == "c0" || 
 getprop("ro.product.device") == "p4notewifi" || getprop("ro.build.product") == "p4notewifi" || 
 getprop("ro.product.device") == "p4notewifiww" || getprop("ro.build.product") == "p4notewifiww" || 
 getprop("ro.product.device") == "n8013" || getprop("ro.build.product") == "n8013" || 
 getprop("ro.product.device") == "GT-N8013" || getprop("ro.build.product") == "GT-N8013" || 
 getprop("ro.product.device") == "n8010" || getprop("ro.build.product") == "n8010" || 
 getprop("ro.product.device") == "GT-N8010" || getprop("ro.build.product") == "GT-N8010" || 
 getprop("ro.product.device") == "n801x" || getprop("ro.build.product") == "n801x" || 
 getprop("ro.product.device") == "GT-N801x" || getprop("ro.build.product") == "GT-N801x") || abort("This package is for \"c0,p4notewifi,p4notewifiww,n8013,GT-N8013,n8010,GT-N8010,n801x,GT-N801x\" devices
this is a \"" + getprop("ro.product.device") + "\".");

But my point was that other ROMs will not include n801x in their asserts so it could cause issues. Although it turns out that TWRP reads the properties from /default.prop not /system/build.prop, so it might not matter as long as the recovery is built with a real model name. But I still think it's better to have a real device model because we don't know where someone might decide to have some logic based on these properties and it might cause unknown issues down the line. In particular Play store/device management and the like could get confusing.

Patch-set: 4