Update patch set 4
Patch Set 4:

I honestly prefer #1 out of those two, it's not just other ROMs but things like Play store may be confused by having a fake device. But I don't really see a problem with having a build for a "virtual" device that can be flashed on either but identifies as one of them for the purposes of acting normal to outside consumers. To be honest my stronger preference would be to keep it as n8013 and only change the asserts so that it can be installed on n8010 as well, so that we have no fake device nonsense but people on n8010 don't feel left out. I feel the "cleanest" solution is two separate builds but it's kind of wasteful and pointless.

Patch-set: 4