Update patch set 4
Patch Set 4:

Just for the record, the Tab 2s are set up so that there's the p5100 build and the p5110 build which can be used on both p5110 and p5113. This keeps compatibility with various other things, and so, I'm personally in favour of setting it up like that. However, I don't have this device, so it's not my call.

Patch-set: 4
Reviewer: Gerrit User 1000032 <1000032@56841c99-7c94-43d4-bf0b-cb4707e66d55>
Label: Code-Review=0