Update patch set 1
Patch Set 1:

Not sure about the dmesg thing (might need to set the loglevel in init.rc or something) but I switched the debugging statements to ERRORs and it does enter the right if blocks. But the problem seems to be that as you said, those properties are already in the build.prop so they can't be overridden. I think we need one or both of those changes from android_build to actually get this to work.

Patch-set: 1
Label: Code-Review=0