Update patch set 3
Patch Set 3: Code-Review+1

Requires https://gerrit.omnirom.org/#/c/5361/

(You'll need to do a clobber after picking that and this for things to work)

We may eventually want to add the build fingerprint to this, but not sure if we want that now.  Gonna have to update the fingerprint soon.  :)

Patch-set: 3
Label: Code-Review=+1