Update patch set 3
Patch Set 3:

If n8010 starts identifying as n8010, we need to do something to fix audio since it uses ro.product.device to determine which conf file to load in audio_hw.c. This is probably the easiest solution, just copy the same conf file twice for each model. We could also use a different property, such as ro.omni.device, but I don't know if this would make it more annoying to keep this audio HAL in sync with whatever else is using it (if that's even a concern). Also, using ro.omni.device would make this audio HAL specific to Omni which seems undesirable (just one extra thing to change when porting a device tree to a different ROM).

Or honestly, I don't really know why it needs to read from a property at all if this audio HAL is actually specific to the n80xx-common repo, if it's going to use the same file for all n80xx-common devices, why not just hardcode the config file name. If the HAL was in smdk4412-common that might make more sense, but even then as long as each device copies its own version of the conf file to the same path I don't see the point of reading ro.product.device at all.

Patch-set: 3